# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Educational Excel-like API",
        "title": "ITSUXEL",
        "contact": {
            "email": "iostapovychweb@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/api/v1",
    "paths": {
        "/table/{sheet}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "returns cell value",
                "parameters": [
                    {
                        "type": "string",
                        "description": "sheet name",
                        "name": "sheet",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/api.ValueResponse"
                            }
                        }
                    }
                }
            }
        },
        "/table/{sheet}/{cell}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "returns cell value",
                "parameters": [
                    {
                        "type": "string",
                        "description": "sheet name",
                        "name": "sheet",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "cell address",
                        "name": "cell",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/api.ValueResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Add a new pet to the store",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "adds value to cell",
                "parameters": [
                    {
                        "type": "string",
                        "description": "sheet name",
                        "name": "sheet",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "cell address",
                        "name": "cell",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Set Cell Value",
                        "name": "json",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.SetCellValueBody"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.SetCellValueBody": {
            "type": "object",
            "required": [
                "value"
            ],
            "properties": {
                "value": {
                    "type": "string"
                }
            }
        },
        "api.ValueResponse": {
            "type": "object",
            "properties": {
                "result": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        }
    }
}